+++ crates/swc_ecma_minifier/src/compress/optimize/iife.rs
@@ -137,6 +137,7 @@ where
     ///     })(x);
     /// })(7);
     /// ```
+    #[cfg_attr(feature = "debug", tracing::instrument(skip(self, e)))]
     pub(super) fn inline_args_of_iife(&mut self, e: &mut CallExpr) {
         if self.options.inline == 0 {
             return;
@@ -250,6 +251,7 @@ where
         }
     }
 
+    #[cfg_attr(feature = "debug", tracing::instrument(skip(self, n, vars)))]
     pub(super) fn inline_vars_in_node<N>(&mut self, n: &mut N, vars: AHashMap<Id, Box<Expr>>)
     where
         N: VisitMutWith<Self>,
@@ -616,6 +618,9 @@ where
                 .collect::<Vec<_>>();
 
             if !vars.is_empty() {
+                if cfg!(feature = "debug") {
+                    tracing::debug!("iife: Creating variables: {:?}", vars);
+                }
                 self.prepend_stmts.push(Stmt::Decl(Decl::Var(VarDecl {
                     span: DUMMY_SP.apply_mark(self.marks.non_top_level),
                     kind: VarDeclKind::Var,